requirements.txt
anthropic
python-dotenv
pydantic
pydantic-settings
